demo: add an EditText control in puzzle view
IMPORTANT:

This is just a demo, trying to give you a taste about my idea.

I'm not familiar with java and android, so there are many things
needed to tweak, for instance:

1, The program will crash if the user enters anything except a
number into EditText.

2, The program should either give a Toast or AltertDialog when
the user enters a puzzle number which does not exist.

3, the layout need to be done by someone who is really familiar
with Android

4, Maybe remove the SeekBar? change it to a ProgressBar?
Sometimes I'll acidentally touch the SeekBar and jump to another
puzzle.